refactored component imports to reflect new components directory structure
