+++ src/components/cloud-edit/cloud-edit.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Wordcloud from './word-cloud';
+import Wordcloud from '../word-cloud/word-cloud';
 
 export default function CloudEdit(props) {
   return (
