+++ src/components/home-view/home-view.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { connect } from 'react-redux';
 
-import GeneratorForm from './generator-form';
+import GeneratorForm from '../generator-form/generator-form';
 // import LibraryCarousel from './library-carousel';
 
 import { fetchClouds } from '../actions';
